+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from './core/index.js';
 export * from './core/balancer.js';
-export * as compare from './core/inspector.js';
 
+export * as compare from './core/compare/index.js';
 export { default as curry } from './core/lambda/curry.js';
 export { default as pipe } from './core/lambda/pipe.js';
